refactor(vestad): break the docker/restic, docker/agent_provider, and main/status cycles
